fix(nwor): Change accepted_len from int to int64_t for PyTorch compatibility
CRITICAL FIX: PyTorch's boxing layer requires all integer parameters to be
int64_t, not int. The compiler error was:

  "You tried to register a kernel with an unsupported integral input type.
   Please use int64_t instead"

Changed in 3 locations:
- csrc/cache.h: Function declaration
- csrc/cache_kernels.cu: Function implementation
- Schema in torch_bindings.cpp already correct ("int" in schema → int64_t in C++)

This allows the kernel to compile and register correctly with PyTorch.

Satya Nadella status: SAVED ✅

+++ csrc/cache.h
@@ -50,7 +50,7 @@ void commit_staged_kv_cache_flash(
     torch::Tensor& staging_metadata,
     torch::Tensor& key_cache,
     torch::Tensor& value_cache,
-    const int accepted_len);
+    const int64_t accepted_len);
 
 void concat_and_cache_mla(torch::Tensor& kv_c, torch::Tensor& k_pe,
                           torch::Tensor& kv_cache, torch::Tensor& slot_mapping,
